requirements.txt
dash==2.8.1
pandas
gunicorn
dash_bootstrap_components
numpy
plotly.express
dash_bootstrap_templates
os
plotly.graph_objects
